create component for support form
